Removing the old fabs derivations helper so we can use the name for the new file
