+++ dataactbroker/helpers/fabs_derivations_helper.py
@@ -1,657 +1,1143 @@
-import re
 import logging
 
 from datetime import datetime
-from sqlalchemy import func, or_, DATE
 
-from dataactcore.models.domainModels import Zips, CityCode, ZipCity, DUNS
-from dataactcore.models.stagingModels import PublishedAwardFinancialAssistance
 from dataactcore.models.lookups import (ACTION_TYPE_DICT, ASSISTANCE_TYPE_DICT, CORRECTION_DELETE_IND_DICT,
                                         RECORD_TYPE_DICT, BUSINESS_TYPE_DICT, BUSINESS_FUNDS_IND_DICT)
-from dataactcore.utils.business_categories import get_business_categories
+from dataactcore.utils.business_categories import derive_fabs_business_categories
 
 logger = logging.getLogger(__name__)
 
 
-def get_zip_data(sess, zip_five, zip_four):
-    """ Get zip data based on 5-digit or 9-digit zips and the counts of congressional districts associated with them
+def log_derivation(message, submission_id, start_time=None):
+    """ Just logging the time taken to run whatever derivation is being run.
+
+        Args:
+            message: the message to log
+            submission_id: the ID of the submission
+            start_time: If provided, use it to calculate the duration.
+    """
+    log_message = {
+        'message': message,
+        'message_type': 'BrokerDebug',
+        'submission_id': submission_id
+    }
+
+    if start_time:
+        log_message['duration'] = datetime.now() - start_time
+    logger.info(log_message)
+
+
+def derive_total_funding_amount(sess, submission_id):
+    """ Deriving the total funding amounts
 
         Args:
             sess: the current DB session
-            zip_five: the 5-digit zip being checked
-            zip_four: the 4-digit zip being checked
-
-        Returns:
-            A Zips object (if one was found) and a count of the congressional districts in that district as 2 return
-            values
+            submission_id: The ID of the submission derivations are being run for
     """
-    zip_info = None
-    cd_count = 1
+    start_time = datetime.now()
+    log_derivation('Beginning total_funding_amount derivation', submission_id)
 
-    # if we have a 4-digit zip to work with, try using both
-    if zip_four:
-        zip_info = sess.query(Zips).filter_by(zip5=zip_five, zip_last4=zip_four).first()
-    # if we didn't manage to find anything using 9 digits or we don't have 9 digits, try to find one using just 5 digits
-    if not zip_info:
-        zip_info = sess.query(Zips).filter_by(zip5=zip_five).first()
-        # if this is a 5-digit zip, there may be more than one congressional district associated with it
-        cd_count = sess.query(Zips.congressional_district_no.label('cd_count')). \
-            filter_by(zip5=zip_five).distinct().count()
+    query = """
+        UPDATE published_award_financial_assistance
+        SET total_funding_amount = COALESCE(federal_action_obligation, 0) + COALESCE(non_federal_funding_amount, 0)
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D';
+    """
+    sess.execute(query.format(submission_id=submission_id))
 
-    return zip_info, cd_count
+    log_derivation('Completed total_funding_amount derivation', submission_id, start_time)
 
 
-def derive_cfda(obj, cfda_dict, job_id, detached_award_financial_assistance_id):
+def derive_cfda(sess, submission_id):
     """ Deriving cfda title from cfda number using cfda program table.
 
         Args:
-            obj: a dictionary containing the details we need to derive from and to
-            cfda_dict: a dictionary containing data for all CFDA objects keyed by cfda number
-            job_id: the ID of the submission job
-            detached_award_financial_assistance_id: the ID of the submission row
+            sess: the current DB session
+            submission_id: The ID of the submission derivations are being run for
     """
-    obj['cfda_title'] = cfda_dict.get(obj['cfda_number'])
-    if not obj['cfda_title']:
-        logger.error({
-            'message': 'CFDA title not found for CFDA number {}'.format(obj['cfda_number']),
-            'message_type': 'BrokerError',
-            'job_id': job_id,
-            'detached_award_financial_assistance_id': detached_award_financial_assistance_id
-        })
+    # TODO: Put the warning back in somehow, maybe do a distinct select of all empty cfda titles
+    start_time = datetime.now()
+    log_derivation('Beginning cfda_title derivation', submission_id)
+
+    query = """
+        UPDATE published_award_financial_assistance AS pafa
+        SET cfda_title = cfda.program_title
+        FROM cfda_program AS cfda
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+            AND pafa.cfda_number = to_char(cfda.program_number, 'FM00.000');
+    """
+    sess.execute(query.format(submission_id=submission_id))
+
+    log_derivation('Completed cfda_title derivation', submission_id, start_time)
 
 
-def derive_awarding_agency_data(obj, sub_tier_dict, office_dict):
+def derive_awarding_agency_data(sess, submission_id):
     """ Deriving awarding sub tier agency name, awarding agency name, and awarding agency code
 
         Args:
-            obj: a dictionary containing the details we need to derive from and to
-            sub_tier_dict: a dictionary containing all the data for SubTierAgency objects keyed by sub tier code
-            office_dict: a dictionary containing all the data for Office objects keyed by office code
+            sess: the current DB session
+            submission_id: The ID of the submission derivations are being run for
     """
-    obj['awarding_agency_name'] = None
-    obj['awarding_sub_tier_agency_n'] = None
+    start_time = datetime.now()
+    log_derivation('Beginning awarding_agency data derivation', submission_id)
 
-    # If we have an office code and no sub tier code, use the office to derive the sub tier
-    if obj['awarding_office_code'] and not obj['awarding_sub_tier_agency_c']:
-        office = office_dict.get(obj['awarding_office_code'].upper())
-        obj['awarding_sub_tier_agency_c'] = office['sub_tier_code']
+    # Deriving awarding sub tier agency code
+    query = """
+        UPDATE published_award_financial_assistance AS pafa
+        SET awarding_sub_tier_agency_c = office.sub_tier_code
+        FROM office
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+            AND COALESCE(awarding_sub_tier_agency_c, '') = ''
+            AND UPPER(pafa.awarding_office_code) = office.office_code;
+    """
+    sess.execute(query.format(submission_id=submission_id))
 
-    # If we have an office code (provided or derived), we can use that to get the names for the sub and top tiers and
-    # the code for the top tier
-    if obj['awarding_sub_tier_agency_c']:
-        sub_tier = sub_tier_dict.get(obj['awarding_sub_tier_agency_c'].upper())
-        obj['awarding_agency_code'] = sub_tier["frec_code"] if sub_tier["is_frec"] else sub_tier["cgac_code"]
-        obj['awarding_agency_name'] = sub_tier["agency_name"]
-        obj['awarding_sub_tier_agency_n'] = sub_tier["sub_tier_agency_name"]
+    # Deriving awarding agency code/name and sub tier name
+    query = """
+        WITH agency_list AS
+            (SELECT (CASE WHEN sta.is_frec
+                        THEN frec.frec_code
+                        ELSE cgac.cgac_code
+                        END) AS agency_code,
+                (CASE WHEN sta.is_frec
+                    THEN frec.agency_name
+                    ELSE cgac.agency_name
+                    END) AS agency_name,
+                sta.sub_tier_agency_code AS sub_tier_code,
+                sta.sub_tier_agency_name AS sub_tier_name
+            FROM sub_tier_agency AS sta
+                INNER JOIN cgac
+                    ON cgac.cgac_id = sta.cgac_id
+                INNER JOIN frec
+                    ON frec.frec_id = sta.frec_id)
+        UPDATE published_award_financial_assistance AS pafa
+        SET awarding_agency_code = agency_code,
+            awarding_agency_name = agency_name,
+            awarding_sub_tier_agency_n = sub_tier_name
+        FROM agency_list
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+            AND UPPER(awarding_sub_tier_agency_c) = sub_tier_code;
+    """
+    sess.execute(query.format(submission_id=submission_id))
+
+    log_derivation('Completed awarding_agency data derivation', submission_id, start_time)
 
 
-def derive_funding_agency_data(obj, sub_tier_dict, office_dict):
+def derive_funding_agency_data(sess, submission_id):
     """ Deriving funding sub tier agency name, funding agency name, and funding agency code
 
         Args:
-            obj: a dictionary containing the details we need to derive from and to
-            sub_tier_dict: a dictionary containing all the data for SubTierAgency objects keyed by sub tier code
-            office_dict: a dictionary containing all the data for Office objects keyed by office code
+            sess: the current DB session
+            submission_id: The ID of the submission derivations are being run for
     """
-    obj['funding_sub_tier_agency_na'] = None
-    obj['funding_agency_name'] = None
+    start_time = datetime.now()
+    log_derivation('Beginning funding_agency data derivation', submission_id)
 
-    # If we have an office code and no sub tier code, use the office to derive the sub tier
-    if obj['funding_office_code'] and not obj['funding_sub_tier_agency_co']:
-        office = office_dict.get(obj['funding_office_code'].upper())
-        obj['funding_sub_tier_agency_co'] = office['sub_tier_code']
+    # Deriving funding sub tier agency code
+    query = """
+        UPDATE published_award_financial_assistance AS pafa
+        SET funding_sub_tier_agency_co = office.sub_tier_code
+        FROM office
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+            AND COALESCE(funding_sub_tier_agency_co, '') = ''
+            AND UPPER(pafa.funding_office_code) = office.office_code;
+    """
+    sess.execute(query.format(submission_id=submission_id))
 
-    if obj['funding_sub_tier_agency_co']:
-        sub_tier = sub_tier_dict.get(obj['funding_sub_tier_agency_co'].upper())
-        obj['funding_agency_code'] = sub_tier["frec_code"] if sub_tier["is_frec"] else sub_tier["cgac_code"]
-        obj['funding_agency_name'] = sub_tier["agency_name"]
-        obj['funding_sub_tier_agency_na'] = sub_tier["sub_tier_agency_name"]
+    # Deriving funding agency code/name and sub tier name
+    query = """
+        WITH agency_list AS
+            (SELECT (CASE WHEN sta.is_frec
+                        THEN frec.frec_code
+                        ELSE cgac.cgac_code
+                        END) AS agency_code,
+                (CASE WHEN sta.is_frec
+                    THEN frec.agency_name
+                    ELSE cgac.agency_name
+                    END) AS agency_name,
+                sta.sub_tier_agency_code AS sub_tier_code,
+                sta.sub_tier_agency_name AS sub_tier_name
+            FROM sub_tier_agency AS sta
+                INNER JOIN cgac
+                    ON cgac.cgac_id = sta.cgac_id
+                INNER JOIN frec
+                    ON frec.frec_id = sta.frec_id)
+        UPDATE published_award_financial_assistance AS pafa
+        SET funding_agency_code = agency_code,
+            funding_agency_name = agency_name,
+            funding_sub_tier_agency_na = sub_tier_name
+        FROM agency_list
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+            AND UPPER(funding_sub_tier_agency_co) = sub_tier_code;
+    """
+    sess.execute(query.format(submission_id=submission_id))
+
+    log_derivation('Completed funding_agency data derivation', submission_id, start_time)
 
 
-def derive_ppop_state(obj, state_dict):
+def derive_ppop_state(sess, submission_id):
     """ Deriving ppop code and ppop state name
 
         Args:
-            obj: a dictionary containing the details we need to derive from and to
-            state_dict: a dictionary containing all the data for State objects keyed by state code
-
-        Returns:
-            Place of performance code, state code, and state name as 3 return values (all strings or None)
-    """
-    ppop_code = None
-    state_code = None
-    state_name = None
-    if obj['place_of_performance_code']:
-        ppop_code = obj['place_of_performance_code'].upper()
-        if ppop_code == '00*****':
-            state_name = 'Multi-state'
-        elif ppop_code != '00FORGN':
-            state_code = ppop_code[:2]
-            state_name = state_dict.get(state_code)
-
-    obj['place_of_perfor_state_code'] = state_code
-    obj['place_of_perform_state_nam'] = state_name
-
-    return ppop_code, state_code, state_name
-
-
-def derive_ppop_location_data(obj, sess, ppop_code, ppop_state_code, county_dict):
-    """ Deriving place of performance location values from zip4
-
-        Args:
-            obj: a dictionary containing the details we need to derive from and to
             sess: the current DB session
-            ppop_code: place of performance code
-            ppop_state_code: state code from the place of performance code
-            county_dict: a dictionary containing all the data for County objects keyed by state code + county number
+            submission_id: The ID of the submission derivations are being run for
     """
-    if obj['place_of_performance_zip4a'] and obj['place_of_performance_zip4a'].upper() != 'CITY-WIDE':
-        zip_five = obj['place_of_performance_zip4a'][:5]
-        zip_four = None
+    start_time = datetime.now()
+    log_derivation('Beginning place of performance state derivation', submission_id)
 
-        # if zip4 is 9 digits, set the zip_four value to the last 4 digits
-        if len(obj['place_of_performance_zip4a']) > 5:
-            zip_four = obj['place_of_performance_zip4a'][-4:]
-
-        zip_info, cd_count = get_zip_data(sess, zip_five, zip_four)
-
-        # deriving ppop congressional district
-        if not obj['place_of_performance_congr']:
-            if zip_info.congressional_district_no and cd_count == 1:
-                obj['place_of_performance_congr'] = zip_info.congressional_district_no
-            else:
-                obj['place_of_performance_congr'] = '90'
-
-        # deriving PrimaryPlaceOfPerformanceCountyName/Code
-        obj['place_of_perform_county_co'] = zip_info.county_number
-        obj['place_of_perform_county_na'] = county_dict.get(zip_info.state_abbreviation.upper() +
-                                                            zip_info.county_number)
-
-        # deriving PrimaryPlaceOfPerformanceCityName
-        city_info = sess.query(ZipCity).filter_by(zip_code=zip_five).one()
-        obj['place_of_performance_city'] = city_info.city_name
-
-    # if there is no valid ppop zip4, we need to try to derive county/city info from the ppop code
-    elif ppop_code:
-        # if ppop_code is in county format,
-        if re.match('^[A-Z]{2}\*\*\d{3}$', ppop_code):
-            # getting county name
-            county_code = ppop_code[-3:]
-            obj['place_of_perform_county_co'] = county_code
-            obj['place_of_perform_county_na'] = county_dict.get(ppop_state_code + county_code)
-            obj['place_of_performance_city'] = None
-        # if ppop_code is in city format
-        elif re.match('^[A-Z]{2}\d{5}$', ppop_code) and not re.match('^[A-Z]{2}0{5}$', ppop_code):
-            # getting city and county name
-            city_code = ppop_code[-5:]
-            city_info = sess.query(CityCode).filter(CityCode.city_code == city_code,
-                                                    func.upper(CityCode.state_code) == ppop_state_code).first()
-            obj['place_of_performance_city'] = city_info.feature_name
-            obj['place_of_perform_county_co'] = city_info.county_number
-            obj['place_of_perform_county_na'] = city_info.county_name
-
-    # if there's no ppop code, just set them all to None
-    else:
-        obj['place_of_perform_county_co'] = None
-        obj['place_of_perform_county_na'] = None
-        obj['place_of_performance_city'] = None
-
-
-def derive_ppop_scope(obj):
-    """ Deriving place of performance scope values from zip4 and place of performance code
-
-        Args:
-            obj: a dictionary containing the details we need to derive from and to
+    # Deriving office codes for record type not 1
+    query = """
+        UPDATE published_award_financial_assistance AS pafa
+        SET place_of_perfor_state_code = CASE WHEN UPPER(place_of_performance_code) ~ '^[A-Z][A-Z]'
+                                              THEN state_code
+                                              ELSE NULL
+                                         END,
+            place_of_perform_state_nam = CASE WHEN place_of_performance_code = '00*****'
+                                              THEN 'Multi-state'
+                                              ELSE state_name
+                                         END
+        FROM states
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+            AND (UPPER(SUBSTRING(place_of_performance_code, 1, 2)) = state_code
+                    OR place_of_performance_code = '00*****');
     """
-    derived_ppop_scope = None
-    ppop_code = obj['place_of_performance_code'].upper() if obj['place_of_performance_code'] else None
-    ppop_zip = obj['place_of_performance_zip4a'].upper() if obj['place_of_performance_zip4a'] else None
+    sess.execute(query.format(submission_id=submission_id))
 
-    if ppop_zip and (ppop_code and (re.match('^[A-Z]{2}\d{5}$', ppop_code) or re.match('^[A-Z]{2}\d{4}R$', ppop_code))):
-        if ppop_zip.upper() == 'CITY-WIDE':
-            derived_ppop_scope = "City-wide"
-        elif re.match('^\d{5}$', ppop_zip) or re.match('^\d{9}$', ppop_zip) or re.match('^\d{5}-\d{4}$', ppop_zip):
-            derived_ppop_scope = "Single ZIP Code"
-    elif ppop_code:
-        # County
-        if re.match('^[A-Z]{2}\*\*\d{3}$', ppop_code):
-            derived_ppop_scope = "County-wide"
-        # City
-        elif re.match('^[A-Z]{2}\d{5}$', ppop_code) or re.match('^[A-Z]{2}\d{4}R$', ppop_code):
-            derived_ppop_scope = "City-wide"
-        # State
-        elif re.match('^[A-Z]{2}\*{5}$', ppop_code):
-            derived_ppop_scope = "State-wide"
-        # Multi-state
-        elif re.match('^00\*{5}$', ppop_code):
-            derived_ppop_scope = "Multi-state"
-        # Foreign
-        elif ppop_code == '00FORGN':
-            derived_ppop_scope = "Foreign"
-
-    obj['place_of_performance_scope'] = derived_ppop_scope
+    log_derivation('Completed place of performance state derivation', submission_id, start_time)
 
 
-def derive_le_location_data(obj, sess, ppop_code, state_dict, ppop_state_code, ppop_state_name, county_dict):
-    """ Deriving place of performance location values
-
-        Args:
-            obj: a dictionary containing the details we need to derive from and to
-            sess: the current DB session
-            ppop_code: place of performance code
-            state_dict: a dictionary containing all the data for State objects keyed by state code
-            ppop_state_code: state code from the place of performance code
-            ppop_state_name: state name from the code from place of performance code
-            county_dict: a dictionary containing all the data for County objects keyed by state code + county number
-    """
-    # Deriving from zip code (record type is 2 or 3 in this case)
-    if obj['legal_entity_zip5']:
-        # legal entity city data
-        city_info = sess.query(ZipCity).filter_by(zip_code=obj['legal_entity_zip5']).one()
-        obj['legal_entity_city_name'] = city_info.city_name
-
-        zip_data, cd_count = get_zip_data(sess, obj['legal_entity_zip5'], obj['legal_entity_zip_last4'])
-
-        # deriving legal entity congressional district
-        if not obj['legal_entity_congressional']:
-            if zip_data.congressional_district_no and cd_count == 1:
-                obj['legal_entity_congressional'] = zip_data.congressional_district_no
-            else:
-                obj['legal_entity_congressional'] = '90'
-
-        # legal entity county data
-        obj['legal_entity_county_code'] = zip_data.county_number
-        obj['legal_entity_county_name'] = county_dict.get(zip_data.state_abbreviation.upper() + zip_data.county_number)
-
-        # legal entity state data
-        obj['legal_entity_state_code'] = zip_data.state_abbreviation
-        obj['legal_entity_state_name'] = state_dict.get(zip_data.state_abbreviation.upper())
-
-    # deriving legal entity stuff that's based on record type of 1
-    # (ppop code must be in the format XX**###, XX*****, 00FORGN for these)
-    if obj['record_type'] == 1:
-
-        county_wide_pattern = re.compile("^[a-zA-Z]{2}\*{2}\d{3}$")
-        state_wide_pattern = re.compile("^[a-zA-Z]{2}\*{5}$")
-
-        obj['legal_entity_congressional'] = None
-
-        if county_wide_pattern.match(ppop_code):
-            # legal entity county data
-            county_code = ppop_code[-3:]
-            obj['legal_entity_county_code'] = county_code
-            obj['legal_entity_county_name'] = county_dict.get(ppop_state_code + county_code)
-
-        if county_wide_pattern.match(ppop_code) or state_wide_pattern.match(ppop_code):
-            # legal entity state data
-            obj['legal_entity_state_code'] = ppop_state_code
-            obj['legal_entity_state_name'] = ppop_state_name
-
-        # legal entity cd data
-        if not obj['legal_entity_congressional'] and (county_wide_pattern.match(ppop_code) or
-                                                      state_wide_pattern.match(ppop_code)):
-            obj['legal_entity_congressional'] = obj['place_of_performance_congr']
-
-
-def derive_office_data(obj, office_dict, sess):
-    """ Deriving office data
-
-        Args:
-            obj: a dictionary containing the details we need to derive from and to
-            office_dict: a dictionary containing all the data for Office objects keyed by office code
-            sess: the current DB session
-    """
-    # If we don't have an awarding office code, we need to copy it from the earliest transaction of that award
-    if not obj['awarding_office_code'] or not obj['funding_office_code']:
-        first_transaction = None
-        pafa = PublishedAwardFinancialAssistance
-        awarding_sub_code = obj['awarding_sub_tier_agency_c'].upper() if obj['awarding_sub_tier_agency_c'] else None
-        if obj['record_type'] == 1:
-            uri = obj['uri'].upper() if obj['uri'] else None
-            # Get the minimum action date for this uri/AwardingSubTierCode combo
-            min_action_date = sess.query(func.min(pafa.action_date).label("min_date")). \
-                filter(func.upper(pafa.uri) == uri, func.upper(pafa.awarding_sub_tier_agency_c) == awarding_sub_code,
-                       pafa.is_active.is_(True), pafa.record_type == 1).one()
-            # If we have a minimum action date, get the office codes for the first entry that matches it
-            if min_action_date.min_date:
-                first_transaction = sess.query(pafa.awarding_office_code, pafa.funding_office_code,
-                                               pafa.award_modification_amendme).\
-                    filter(func.upper(pafa.uri) == uri, pafa.is_active.is_(True),
-                           func.upper(pafa.awarding_sub_tier_agency_c) == awarding_sub_code,
-                           func.cast_as_date(pafa.action_date) == min_action_date.min_date,
-                           pafa.record_type == 1).first()
-        else:
-            fain = obj['fain'].upper() if obj['fain'] else None
-            # Get the minimum action date for this fain/AwardingSubTierCode combo
-            min_action_date = sess.query(func.min(func.cast(pafa.action_date, DATE)).label("min_date")).\
-                filter(func.upper(pafa.fain) == fain, func.upper(pafa.awarding_sub_tier_agency_c) == awarding_sub_code,
-                       pafa.is_active.is_(True), pafa.record_type != 1).one()
-            # If we have a minimum action date, get the office codes for the first entry that matches it
-            if min_action_date.min_date:
-                first_transaction = sess.query(pafa.awarding_office_code, pafa.funding_office_code,
-                                               pafa.award_modification_amendme).\
-                    filter(func.upper(pafa.fain) == fain, pafa.is_active.is_(True),
-                           func.upper(pafa.awarding_sub_tier_agency_c) == awarding_sub_code,
-                           func.cast_as_date(pafa.action_date) == min_action_date.min_date,
-                           pafa.record_type != 1).first()
-
-        # If we managed to find a transaction, copy the office codes into it. Don't copy if the mod is the same because
-        # we don't want to auto-fill the base record for an award.
-        if first_transaction and first_transaction.award_modification_amendme != obj['award_modification_amendme']:
-            # No need to copy if new code isn't blank or old code is
-            if not obj['awarding_office_code'] and first_transaction.awarding_office_code:
-                # Make sure the code we're copying is a valid awarding office code
-                award_office = office_dict.get(first_transaction.awarding_office_code.upper())
-                if award_office and award_office['financial_assistance_awards_office']:
-                    obj['awarding_office_code'] = first_transaction.awarding_office_code
-            if not obj['funding_office_code'] and first_transaction.funding_office_code:
-                # Make sure the code we're copying is a valid funding office code
-                fund_office = office_dict.get(first_transaction.funding_office_code.upper())
-                if fund_office and fund_office['funding_office']:
-                    obj['funding_office_code'] = first_transaction.funding_office_code
-
-    # Deriving awarding_office_name based off awarding_office_code
-    awarding_code = obj['awarding_office_code'].upper() if obj['awarding_office_code'] else None
-    awarding_office_data = office_dict.get(awarding_code)
-    if awarding_office_data:
-        obj['awarding_office_name'] = awarding_office_data['office_name']
-    else:
-        obj['awarding_office_name'] = None
-
-    # Deriving funding_office_name based off funding_office_code
-    funding_code = obj['funding_office_code'].upper() if obj['funding_office_code'] else None
-    funding_office_data = office_dict.get(funding_code)
-    if funding_office_data:
-        obj['funding_office_name'] = funding_office_data['office_name']
-    else:
-        obj['funding_office_name'] = None
-
-
-def derive_le_city_code(obj, sess):
-    """ Deriving legal entity city code
-
-        Args:
-            obj: a dictionary containing the details we need to derive from and to
-            sess: the current DB session
-    """
-    if obj['legal_entity_city_name'] and obj['legal_entity_state_code']:
-        city_code = sess.query(CityCode).\
-            filter(func.upper(CityCode.feature_name) == func.upper(obj['legal_entity_city_name'].strip()),
-                   func.upper(CityCode.state_code) == func.upper(obj['legal_entity_state_code'].strip())).first()
-        if city_code:
-            obj['legal_entity_city_code'] = city_code.city_code
-        else:
-            obj['legal_entity_city_code'] = None
-
-
-def derive_ppop_country_name(obj, country_dict):
-    """ Deriving place of performance country name
-
-        Args:
-            obj: a dictionary containing the details we need to derive from and to
-            country_dict: a dictionary containing all the data for Country objects keyed by country code
-    """
-    if obj['place_of_perform_country_c']:
-        obj['place_of_perform_country_n'] = country_dict.get(obj['place_of_perform_country_c'].upper())
-
-
-def derive_le_country_name(obj, country_dict):
-    """ Deriving legal entity country name
-
-        Args:
-            obj: a dictionary containing the details we need to derive from and to
-            country_dict: a dictionary containing all the data for Country objects keyed by country code
-    """
-    if obj['legal_entity_country_code']:
-        obj['legal_entity_country_name'] = country_dict.get(obj['legal_entity_country_code'].upper())
-
-
-def derive_ppop_code(obj):
-    """ Deriving place of performance code for PII-redacted records
-
-        Args:
-            obj: a dictionary containing the details we need to derive from and to
-    """
-    if obj['record_type'] == 3:
-        if obj['legal_entity_country_code'].upper() == 'USA' and obj['legal_entity_state_code']:
-            ppop_code = obj['legal_entity_state_code'].upper()
-            if obj['legal_entity_city_code']:
-                obj['place_of_performance_code'] = ppop_code + obj['legal_entity_city_code']
-            else:
-                obj['place_of_performance_code'] = ppop_code + '00000'
-        elif obj['legal_entity_country_code'].upper() != 'USA':
-            obj['place_of_performance_code'] = '00FORGN'
-
-
-def derive_pii_redacted_ppop_data(obj):
-    """ Deriving ppop location data for PII-redacted records
-
-        Args:
-            obj: a dictionary containing the details we need to derive from and to
-    """
-    if obj['record_type'] == 3:
-        # These are the same no matter where the country code is
-        obj['place_of_perform_country_c'] = obj['legal_entity_country_code']
-        obj['place_of_perform_country_n'] = obj['legal_entity_country_name']
-        # Deriving data for domestic recipients
-        if obj['legal_entity_country_code'].upper() == 'USA':
-            obj['place_of_performance_city'] = obj['legal_entity_city_name']
-            obj['place_of_perform_county_co'] = obj['legal_entity_county_code']
-            obj['place_of_perform_county_na'] = obj['legal_entity_county_name']
-            obj['place_of_perfor_state_code'] = obj['legal_entity_state_code']
-            obj['place_of_perform_state_nam'] = obj['legal_entity_state_name']
-            obj['place_of_performance_zip4a'] = obj['legal_entity_zip5']
-            obj['place_of_performance_congr'] = obj['legal_entity_congressional']
-        else:
-            obj['place_of_performance_city'] = obj['legal_entity_foreign_city']
-            obj['place_of_performance_forei'] = obj['legal_entity_foreign_city']
-
-
-def split_ppop_zip(obj):
+def split_ppop_zip(sess, submission_id):
     """ Splitting ppop zip code into 5 and 4 digit codes for ease of website access
 
         Args:
-            obj: a dictionary containing the details we need to derive from and to
+            sess: the current DB session
+            submission_id: The ID of the submission derivations are being run for
     """
-    if obj['place_of_performance_zip4a'] and re.match('^\d{5}(-?\d{4})?$', obj['place_of_performance_zip4a']):
-        if len(obj['place_of_performance_zip4a']) == 5:
-            obj['place_of_performance_zip5'] = obj['place_of_performance_zip4a'][:5]
-            obj['place_of_perform_zip_last4'] = None
-        else:
-            obj['place_of_performance_zip5'] = obj['place_of_performance_zip4a'][:5]
-            obj['place_of_perform_zip_last4'] = obj['place_of_performance_zip4a'][-4:]
+    start_time = datetime.now()
+    log_derivation('Beginning place of performance zip5 and zip last4 derivation', submission_id)
+
+    query = """
+        UPDATE published_award_financial_assistance AS pafa
+        SET place_of_performance_zip5 = SUBSTRING(place_of_performance_zip4a, 1, 5),
+            place_of_perform_zip_last4 = CASE WHEN LENGTH(place_of_performance_zip4a) = 5
+                                              THEN NULL
+                                              ELSE RIGHT(place_of_performance_zip4a, 4)
+                                         END
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+            AND place_of_performance_zip4a ~ '^\d\d\d\d\d(-?\d\d\d\d)?$';
+    """
+    sess.execute(query.format(submission_id=submission_id))
+
+    log_derivation('Completed place of performance zip5 and zip last4 derivation', submission_id, start_time)
 
 
-def derive_parent_duns(obj, sess):
+def derive_ppop_location_data(sess, submission_id):
+    """ Deriving place of performance location values from zip4
+
+        Args:
+            sess: the current DB session
+            submission_id: The ID of the submission derivations are being run for
+    """
+    start_time = datetime.now()
+    log_derivation('Beginning place of performance location derivation', submission_id)
+
+    # Deriving congressional and county info for records with a 9 digit zip
+    query = """
+        UPDATE published_award_financial_assistance
+        SET place_of_performance_congr = CASE WHEN place_of_performance_congr IS NULL
+                                              THEN congressional_district_no
+                                              ELSE place_of_performance_congr
+                                         END,
+            place_of_perform_county_co = county_number
+        FROM zips
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+            AND place_of_perform_zip_last4 IS NOT NULL
+            AND place_of_perform_zip_last4 = zip_last4
+            AND place_of_performance_zip5 = zip5;
+    """
+    sess.execute(query.format(submission_id=submission_id))
+
+    # Deriving congressional info for multi-district zips
+    query = """
+        WITH all_sub_zips AS
+            (SELECT DISTINCT place_of_performance_zip5
+            FROM published_award_financial_assistance
+            WHERE submission_id = {submission_id}
+                AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+                AND place_of_performance_congr IS NULL),
+        congr_dist AS
+            (SELECT COUNT(DISTINCT congressional_district_no) AS cd_count, zip5
+            FROM zips
+            WHERE EXISTS (
+                SELECT 1
+                FROM all_sub_zips AS asz
+                WHERE zips.zip5 = asz.place_of_performance_zip5)
+            GROUP BY zip5)
+        UPDATE published_award_financial_assistance
+        SET place_of_performance_congr = CASE WHEN cd_count <> 1
+                                              THEN '90'
+                                         END
+        FROM congr_dist
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+            AND place_of_performance_congr IS NULL
+            AND zip5 = place_of_performance_zip5;
+    """
+    sess.execute(query.format(submission_id=submission_id))
+
+    # Deriving congressional info for remaining blanks (with zip code)
+    query = """
+        UPDATE published_award_financial_assistance
+        SET place_of_performance_congr = congressional_district_no
+        FROM zips
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+            AND place_of_performance_zip5 = zip5
+            AND place_of_performance_congr IS NULL;
+    """
+    sess.execute(query.format(submission_id=submission_id))
+
+    # Deriving county code info for remaining blanks (with zip code)
+    query = """
+        UPDATE published_award_financial_assistance
+        SET place_of_perform_county_co = county_number
+        FROM zips
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+            AND place_of_performance_zip5 = zip5
+            AND place_of_perform_county_co IS NULL;
+    """
+    sess.execute(query.format(submission_id=submission_id))
+
+    # Deriving city info for transactions with zips
+    query = """
+        UPDATE published_award_financial_assistance
+        SET place_of_performance_city = city_name
+        FROM zip_city
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+            AND place_of_performance_zip5 IS NOT NULL
+            AND zip_city.zip_code = place_of_performance_zip5;
+    """
+    sess.execute(query.format(submission_id=submission_id))
+
+    # Deriving county code info for transactions with ppop code XX**###
+    query = """
+        UPDATE published_award_financial_assistance
+        SET place_of_perform_county_co = RIGHT(place_of_performance_code, 3)
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+            AND place_of_performance_zip5 IS NULL
+            AND UPPER(place_of_performance_code) ~ '^[A-Z][A-Z]\*\*\d\d\d$';
+    """
+    sess.execute(query.format(submission_id=submission_id))
+
+    # Deriving county/city info for transactions with ppop code XX#####
+    query = """
+        UPDATE published_award_financial_assistance
+        SET place_of_perform_county_co = county_number,
+            place_of_perform_county_na = county_name,
+            place_of_performance_city = feature_name
+        FROM city_code AS cc
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+            AND place_of_performance_zip5 IS NULL
+            AND UPPER(place_of_performance_code) ~ '^[A-Z][A-Z]\d\d\d\d\d$'
+            AND cc.city_code = RIGHT(place_of_performance_code, 5)
+            AND cc.state_code = place_of_perfor_state_code;
+    """
+    sess.execute(query.format(submission_id=submission_id))
+
+    # Deriving remaining county names
+    query = """
+        UPDATE published_award_financial_assistance
+        SET place_of_perform_county_na = county_name
+        FROM county_code AS cc
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+            AND place_of_perform_county_na IS NULL
+            AND place_of_perform_county_co IS NOT NULL
+            AND cc.county_number = place_of_perform_county_co
+            AND cc.state_code = place_of_perfor_state_code;
+    """
+    sess.execute(query.format(submission_id=submission_id))
+
+    log_derivation('Completed place of performance location derivation', submission_id, start_time)
+
+
+def derive_ppop_scope(sess, submission_id):
+    """ Deriving place of performance scope values from zip4 and place of performance code
+
+        Args:
+            sess: the current DB session
+            submission_id: The ID of the submission derivations are being run for
+    """
+    start_time = datetime.now()
+    log_derivation('Beginning place of performance scope derivation', submission_id)
+
+    # When zip is not null
+    query = """
+        UPDATE published_award_financial_assistance AS pafa
+        SET place_of_performance_scope = CASE WHEN UPPER(place_of_performance_zip4a) = 'CITY-WIDE'
+                                              THEN 'City-wide'
+                                              WHEN place_of_performance_zip4a ~ '^\d\d\d\d\d(\-?\d\d\d\d)?$'
+                                              THEN 'Single ZIP Code'
+                                              ELSE NULL
+                                         END
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+            AND COALESCE(place_of_performance_zip4a, '') <> ''
+            AND UPPER(place_of_performance_code) ~ '^[A-Z][A-Z]\d\d\d\d[\dR]$';
+    """
+    sess.execute(query.format(submission_id=submission_id))
+
+    # When zip is null
+    query = """
+        UPDATE published_award_financial_assistance AS pafa
+        SET place_of_performance_scope = CASE WHEN UPPER(place_of_performance_code) ~ '^[A-Z][A-Z]\d\d\d\d[\dR]$'
+                                              THEN 'City-wide'
+                                              WHEN UPPER(place_of_performance_code) ~ '^[A-Z][A-Z]\*\*\d\d\d$'
+                                              THEN 'County-wide'
+                                              WHEN UPPER(place_of_performance_code) ~ '^[A-Z][A-Z]\*\*\*\*\*$'
+                                              THEN 'State-wide'
+                                              WHEN UPPER(place_of_performance_code) ~ '^00\*\*\*\*\*$'
+                                              THEN 'Multi-state'
+                                              WHEN UPPER(place_of_performance_code) ~ '^00FORGN$'
+                                              THEN 'Foreign'
+                                              ELSE NULL
+                                         END
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+            AND COALESCE(place_of_performance_zip4a, '') = '';
+    """
+    sess.execute(query.format(submission_id=submission_id))
+
+    log_derivation('Completed place of performance scope derivation', submission_id, start_time)
+
+
+def derive_le_location_data(sess, submission_id):
+    """ Deriving place of performance location values
+
+        Args:
+            sess: the current DB session
+            submission_id: The ID of the submission derivations are being run for
+    """
+    start_time = datetime.now()
+    log_derivation('Beginning legal entity location derivation', submission_id)
+
+    # Deriving congressional, county, and state info for records with a 9 digit zip
+    query = """
+        UPDATE published_award_financial_assistance
+        SET legal_entity_congressional = CASE WHEN legal_entity_congressional IS NULL
+                                              THEN congressional_district_no
+                                              ELSE legal_entity_congressional
+                                         END,
+            legal_entity_county_code = county_number,
+            legal_entity_state_code = state_abbreviation
+        FROM zips
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+            AND legal_entity_zip_last4 IS NOT NULL
+            AND legal_entity_zip_last4 = zip_last4
+            AND legal_entity_zip5 = zip5;
+    """
+    sess.execute(query.format(submission_id=submission_id))
+
+    # Deriving congressional info for multi-district zips
+    query = """
+        WITH all_sub_zips AS
+            (SELECT DISTINCT legal_entity_zip5
+            FROM published_award_financial_assistance
+            WHERE submission_id = {submission_id}
+                AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+                AND legal_entity_congressional IS NULL),
+        congr_dist AS
+            (SELECT COUNT(DISTINCT congressional_district_no) AS cd_count, zip5
+            FROM zips
+            WHERE EXISTS (
+                SELECT 1
+                FROM all_sub_zips AS asz
+                WHERE zips.zip5 = asz.legal_entity_zip5)
+            GROUP BY zip5)
+        UPDATE published_award_financial_assistance
+        SET legal_entity_congressional = CASE WHEN cd_count <> 1
+                                              THEN '90'
+                                         END
+        FROM congr_dist
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+            AND legal_entity_congressional IS NULL
+            AND zip5 = legal_entity_zip5;
+    """
+    sess.execute(query.format(submission_id=submission_id))
+
+    # Deriving congressional info for remaining blanks (with zip code)
+    query = """
+        UPDATE published_award_financial_assistance
+        SET legal_entity_congressional = congressional_district_no
+        FROM zips
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+            AND legal_entity_zip5 = zip5
+            AND legal_entity_congressional IS NULL;
+    """
+    sess.execute(query.format(submission_id=submission_id))
+
+    # Deriving county and state code info for remaining blanks (with zip code)
+    query = """
+        UPDATE published_award_financial_assistance
+        SET legal_entity_county_code = county_number,
+            legal_entity_state_code = state_abbreviation
+        FROM zips
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+            AND legal_entity_zip5 = zip5
+            AND legal_entity_county_code IS NULL;
+    """
+    sess.execute(query.format(submission_id=submission_id))
+
+    # Deriving county names for records with zips (type 2 and 3)
+    query = """
+        UPDATE published_award_financial_assistance
+        SET legal_entity_county_name = county_name
+        FROM county_code AS cc
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+            AND legal_entity_zip5 IS NOT NULL
+            AND cc.county_number = legal_entity_county_code
+            AND cc.state_code = legal_entity_state_code;
+    """
+    sess.execute(query.format(submission_id=submission_id))
+
+    # Deriving state names for records with zips (type 2 and 3)
+    query = """
+        UPDATE published_award_financial_assistance
+        SET legal_entity_state_name = state_name
+        FROM states
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+            AND legal_entity_zip5 IS NOT NULL
+            AND states.state_code = legal_entity_state_code;
+    """
+    sess.execute(query.format(submission_id=submission_id))
+
+    # Deriving city info for records with zips (type 2 and 3)
+    query = """
+        UPDATE published_award_financial_assistance
+        SET legal_entity_city_name = city_name
+        FROM zip_city
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+            AND legal_entity_zip5 IS NOT NULL
+            AND zip_city.zip_code = legal_entity_zip5;
+    """
+    sess.execute(query.format(submission_id=submission_id))
+
+    # Deriving county, state, and congressional info for county format ppop codes in record type 1
+    query = """
+        UPDATE published_award_financial_assistance
+        SET legal_entity_county_code = place_of_perform_county_co,
+            legal_entity_county_name = place_of_perform_county_na,
+            legal_entity_state_code = place_of_perfor_state_code,
+            legal_entity_state_name = place_of_perform_state_nam,
+            legal_entity_congressional = place_of_performance_congr
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+            AND record_type = 1
+            AND UPPER(place_of_performance_code) ~ '^[A-Z][A-Z]\*\*\d\d\d$';
+    """
+    sess.execute(query.format(submission_id=submission_id))
+
+    # Deriving county, state, and congressional info for state format ppop codes in record type 1
+    query = """
+        UPDATE published_award_financial_assistance
+        SET legal_entity_state_code = place_of_perfor_state_code,
+            legal_entity_state_name = place_of_perform_state_nam,
+            legal_entity_congressional = place_of_performance_congr
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+            AND record_type = 1
+            AND UPPER(place_of_performance_code) ~ '^[A-Z][A-Z]\*\*\*\*\*$';
+    """
+    sess.execute(query.format(submission_id=submission_id))
+
+    log_derivation('Completed legal entity location derivation', submission_id, start_time)
+
+
+def derive_office_data(sess, submission_id):
+    """ Deriving office data
+
+        Args:
+            sess: the current DB session
+            submission_id: The ID of the submission derivations are being run for
+    """
+    start_time = datetime.now()
+    log_derivation('Beginning office data derivation', submission_id)
+
+    # Deriving office codes for record type not 1
+    query = """
+        WITH awards AS
+            (SELECT DISTINCT UPPER(fain) AS upper_fain,
+                UPPER(awarding_sub_tier_agency_c) AS upper_sub_tier
+            FROM published_award_financial_assistance
+            WHERE submission_id = {submission_id}
+                AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+                AND record_type <> '1'),
+        min_date AS
+            (SELECT CAST(MIN(action_date) AS DATE) AS min_date,
+                UPPER(fain) AS upper_fain,
+                UPPER(awarding_sub_tier_agency_c) AS upper_sub_tier
+            FROM published_award_financial_assistance AS pafa
+            WHERE is_active IS TRUE
+                AND record_type <> '1'
+                AND submission_id <> {submission_id}
+                AND EXISTS (
+                    SELECT 1
+                    FROM awards
+                    WHERE upper_fain = UPPER(fain)
+                        AND upper_sub_tier = UPPER(awarding_sub_tier_agency_c)
+                )
+            GROUP BY UPPER(fain), UPPER(awarding_sub_tier_agency_c)),
+        office_info AS
+            (SELECT awarding_office_code AS awarding_office_code,
+                funding_office_code AS funding_office_code,
+                award_modification_amendme,
+                UPPER(fain) AS upper_fain,
+                UPPER(awarding_sub_tier_agency_c) AS upper_sub_tier
+            FROM published_award_financial_assistance AS pafa
+            WHERE is_active IS TRUE
+                AND record_type <> '1'
+                AND EXISTS (
+                    SELECT 1
+                    FROM min_date AS md
+                    WHERE upper_fain = UPPER(fain)
+                        AND upper_sub_tier = UPPER(awarding_sub_tier_agency_c)
+                        AND CAST(pafa.action_date AS DATE) = min_date
+                )),
+        filtered_offices AS
+            (SELECT award_modification_amendme,
+                upper_fain,
+                upper_sub_tier,
+                oi.awarding_office_code AS awarding_office_code,
+                oi.funding_office_code AS funding_office_code
+            FROM office_info AS oi
+            LEFT JOIN office AS aw_office
+                ON aw_office.office_code = UPPER(oi.awarding_office_code)
+                AND aw_office.financial_assistance_awards_office IS TRUE
+            LEFT JOIN office AS fund_office
+                ON fund_office.office_code = UPPER(oi.funding_office_code)
+                AND (fund_office.contract_funding_office IS TRUE
+                    OR fund_office.financial_assistance_funding_office IS TRUE))
+        UPDATE published_award_financial_assistance AS pafa
+        SET awarding_office_code = CASE WHEN pafa.awarding_office_code IS NULL
+                                        THEN fo.awarding_office_code
+                                        ELSE pafa.awarding_office_code
+                                   END,
+            funding_office_code = CASE WHEN pafa.funding_office_code IS NULL
+                                       THEN fo.funding_office_code
+                                       ELSE pafa.funding_office_code
+                                  END
+        FROM filtered_offices AS fo
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+            AND COALESCE(pafa.award_modification_amendme, '') <> COALESCE(fo.award_modification_amendme, '')
+            AND upper_fain = fain
+            AND upper_sub_tier = UPPER(awarding_sub_tier_agency_c)
+            AND record_type <> '1';
+    """
+    sess.execute(query.format(submission_id=submission_id))
+
+    # Deriving office codes for record type 1
+    query = """
+        WITH awards AS
+            (SELECT DISTINCT UPPER(uri) AS upper_uri,
+            UPPER(awarding_sub_tier_agency_c) AS upper_sub_tier
+            FROM published_award_financial_assistance
+            WHERE submission_id = {submission_id}
+                AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+                AND record_type = '1'),
+        min_date AS
+            (SELECT CAST(MIN(action_date) AS DATE) AS min_date,
+            UPPER(uri) AS upper_uri, UPPER(awarding_sub_tier_agency_c) AS upper_sub_tier
+            FROM published_award_financial_assistance AS pafa
+            WHERE is_active IS TRUE
+                AND record_type = '1'
+                AND submission_id <> {submission_id}
+                AND EXISTS (
+                    SELECT 1
+                    FROM awards
+                    WHERE upper_uri = UPPER(uri)
+                        AND upper_sub_tier = UPPER(awarding_sub_tier_agency_c)
+                )
+            GROUP BY UPPER(uri), UPPER(awarding_sub_tier_agency_c)),
+        office_info AS
+            (SELECT awarding_office_code AS awarding_office_code,
+                funding_office_code AS funding_office_code,
+                award_modification_amendme,
+                UPPER(uri) AS upper_uri,
+                UPPER(awarding_sub_tier_agency_c) AS upper_sub_tier
+            FROM published_award_financial_assistance AS pafa
+            WHERE is_active IS TRUE
+                AND record_type = '1'
+                AND EXISTS (
+                    SELECT 1
+                    FROM min_date AS md
+                    WHERE upper_uri = UPPER(uri)
+                        AND upper_sub_tier = UPPER(awarding_sub_tier_agency_c)
+                        AND CAST(pafa.action_date AS DATE) = min_date
+                )),
+        filtered_offices AS
+            (SELECT award_modification_amendme,
+                upper_uri,
+                upper_sub_tier,
+                oi.awarding_office_code AS awarding_office_code,
+                oi.funding_office_code AS funding_office_code
+            FROM office_info AS oi
+            LEFT JOIN office AS aw_office
+                ON aw_office.office_code = UPPER(oi.awarding_office_code)
+                AND aw_office.financial_assistance_awards_office IS TRUE
+            LEFT JOIN office AS fund_office
+                ON fund_office.office_code = UPPER(oi.funding_office_code)
+                AND (fund_office.contract_funding_office IS TRUE
+                    OR fund_office.financial_assistance_funding_office IS TRUE))
+        UPDATE published_award_financial_assistance AS pafa
+        SET awarding_office_code = CASE WHEN pafa.awarding_office_code IS NULL
+                                        THEN fo.awarding_office_code
+                                        ELSE pafa.awarding_office_code
+                                   END,
+            funding_office_code = CASE WHEN pafa.funding_office_code IS NULL
+                                       THEN fo.funding_office_code
+                                       ELSE pafa.funding_office_code
+                                  END
+        FROM filtered_offices AS fo
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+            AND COALESCE(pafa.award_modification_amendme, '') <> COALESCE(fo.award_modification_amendme, '')
+            AND upper_uri = uri
+            AND upper_sub_tier = UPPER(awarding_sub_tier_agency_c)
+            AND record_type = '1';
+    """
+    sess.execute(query.format(submission_id=submission_id))
+
+    # Deriving awarding office name
+    query = """
+        UPDATE published_award_financial_assistance
+        SET awarding_office_name = office_name
+        FROM office
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+            AND office_code = UPPER(awarding_office_code);
+    """
+    sess.execute(query.format(submission_id=submission_id))
+
+    # Deriving funding office name
+    query = """
+        UPDATE published_award_financial_assistance
+        SET funding_office_name = office_name
+        FROM office
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+            AND office_code = UPPER(funding_office_code);
+    """
+    sess.execute(query.format(submission_id=submission_id))
+
+    log_derivation('Completed office data derivation', submission_id, start_time)
+
+
+def derive_le_city_code(sess, submission_id):
+    """ Deriving legal entity city code
+
+        Args:
+            sess: the current DB session
+            submission_id: The ID of the submission derivations are being run for
+    """
+    start_time = datetime.now()
+    log_derivation('Beginning legal entity city code derivation', submission_id)
+
+    query = """
+        UPDATE published_award_financial_assistance AS pafa
+        SET legal_entity_city_code = city_code
+        FROM city_code
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+            AND UPPER(TRIM(legal_entity_city_name)) = UPPER(feature_name)
+            AND UPPER(TRIM(legal_entity_state_code)) = UPPER(state_code);
+    """
+    sess.execute(query.format(submission_id=submission_id))
+
+    log_derivation('Completed legal entity city code derivation', submission_id, start_time)
+
+
+def derive_ppop_country_name(sess, submission_id):
+    """ Deriving place of performance country name
+
+        Args:
+            sess: the current DB session
+            submission_id: The ID of the submission derivations are being run for
+    """
+    start_time = datetime.now()
+    log_derivation('Beginning place of performance country name derivation', submission_id)
+
+    query = """
+        UPDATE published_award_financial_assistance AS pafa
+        SET place_of_perform_country_n = country_name
+        FROM country_code
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+            AND country_code.country_code = UPPER(place_of_perform_country_c);
+    """
+    sess.execute(query.format(submission_id=submission_id))
+
+    log_derivation('Completed place of performance country name derivation', submission_id, start_time)
+
+
+def derive_le_country_name(sess, submission_id):
+    """ Deriving legal entity country name
+
+        Args:
+            sess: the current DB session
+            submission_id: The ID of the submission derivations are being run for
+    """
+    start_time = datetime.now()
+    log_derivation('Beginning legal entity country name derivation', submission_id)
+
+    query = """
+        UPDATE published_award_financial_assistance AS pafa
+        SET legal_entity_country_name = country_name
+        FROM country_code
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+            AND country_code.country_code = UPPER(legal_entity_country_code);
+    """
+    sess.execute(query.format(submission_id=submission_id))
+
+    log_derivation('Completed legal entity country name derivation', submission_id, start_time)
+
+
+def derive_pii_redacted_ppop_data(sess, submission_id):
+    """ Deriving ppop code and location data for PII-redacted records
+
+        Args:
+            sess: the current DB session
+            submission_id: The ID of the submission derivations are being run for
+    """
+    start_time = datetime.now()
+    log_derivation('Beginning PII redacted information derivation', submission_id)
+
+    # Deriving information for USA records
+    query = """
+        UPDATE published_award_financial_assistance AS pafa
+        SET place_of_performance_code = CASE WHEN legal_entity_state_code IS NOT NULL
+                                            THEN CASE WHEN legal_entity_city_code IS NOT NULL
+                                                    THEN UPPER(legal_entity_state_code) || legal_entity_city_code
+                                                    ELSE UPPER(legal_entity_state_code) || '00000'
+                                                    END
+                                            ELSE NULL
+                                            END,
+            place_of_perform_country_c = legal_entity_country_code,
+            place_of_perform_country_n = legal_entity_country_name,
+            place_of_performance_city = legal_entity_city_name,
+            place_of_perform_county_co = legal_entity_county_code,
+            place_of_perform_county_na = legal_entity_county_name,
+            place_of_perfor_state_code = legal_entity_state_code,
+            place_of_perform_state_nam = legal_entity_state_name,
+            place_of_performance_zip4a = legal_entity_zip5,
+            place_of_performance_zip5 = legal_entity_zip5,
+            place_of_performance_congr = legal_entity_congressional
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+            AND record_type = 3
+            AND UPPER(legal_entity_country_code) = 'USA';
+    """
+    sess.execute(query.format(submission_id=submission_id))
+
+    # Deriving information for non-USA records
+    query = """
+        UPDATE published_award_financial_assistance AS pafa
+        SET place_of_performance_code = '00FORGN',
+            place_of_perform_country_c = legal_entity_country_code,
+            place_of_perform_country_n = legal_entity_country_name,
+            place_of_performance_city = legal_entity_foreign_city,
+            place_of_performance_forei = legal_entity_foreign_city
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+            AND record_type = 3
+            AND UPPER(legal_entity_country_code) <> 'USA';
+    """
+    sess.execute(query.format(submission_id=submission_id))
+
+    log_derivation('Completed PII redacted information derivation', submission_id, start_time)
+
+
+def derive_parent_duns(sess, submission_id):
     """ Deriving parent DUNS name and number from SAM API
 
         Args:
-            obj: a dictionary containing the details we need to derive from and to
             sess: the current DB session
+            submission_id: The ID of the submission derivations are being run for
     """
-    if obj['awardee_or_recipient_uniqu']:
-        duns_data = sess.query(DUNS).\
-            filter_by(awardee_or_recipient_uniqu=obj['awardee_or_recipient_uniqu']).\
-            filter(or_(DUNS.ultimate_parent_unique_ide.isnot(None), DUNS.ultimate_parent_unique_ide.isnot(None))).\
-            first()
+    start_time = datetime.now()
+    log_derivation('Beginning parent DUNS derivation', submission_id)
 
-        if duns_data:
-            obj['ultimate_parent_legal_enti'] = duns_data.ultimate_parent_legal_enti
-            obj['ultimate_parent_unique_ide'] = duns_data.ultimate_parent_unique_ide
-        else:
-            obj['ultimate_parent_legal_enti'] = None
-            obj['ultimate_parent_unique_ide'] = None
+    query = """
+        UPDATE published_award_financial_assistance AS pafa
+        SET ultimate_parent_legal_enti = duns.ultimate_parent_legal_enti,
+            ultimate_parent_unique_ide = duns.ultimate_parent_unique_ide
+        FROM duns
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+            AND pafa.awardee_or_recipient_uniqu = duns.awardee_or_recipient_uniqu
+            AND (duns.ultimate_parent_legal_enti IS NOT NULL
+                OR duns.ultimate_parent_unique_ide IS NOT NULL);
+    """
+    sess.execute(query.format(submission_id=submission_id))
+
+    log_derivation('Completed parent DUNS derivation', submission_id, start_time)
 
 
-def derive_executive_compensation(obj, exec_comp_dict):
+def derive_executive_compensation(sess, submission_id):
     """ Deriving Executive Compensation information from DUNS.
 
         Args:
-            obj: a dictionary containing the details we need to derive from and to
-            exec_comp_dict: a dictionary containing all the data for Executive Compensation data keyed by DUNS number
+            sess: the current DB session
+            submission_id: The ID of the submission derivations are being run for
     """
-    if obj['awardee_or_recipient_uniqu'] and obj['awardee_or_recipient_uniqu'] in exec_comp_dict.keys():
-        exec_comp = exec_comp_dict[obj['awardee_or_recipient_uniqu']]
+    start_time = datetime.now()
+    log_derivation('Beginning executive compensation derivation', submission_id)
 
-        for i in range(1, 6):
-            obj['high_comp_officer{}_full_na'.format(i)] = exec_comp['officer{}_name'.format(i)]
-            obj['high_comp_officer{}_amount'.format(i)] = exec_comp['officer{}_amt'.format(i)]
-    else:
-        for i in range(1, 6):
-            obj['high_comp_officer{}_full_na'.format(i)] = None
-            obj['high_comp_officer{}_amount'.format(i)] = None
+    query = """
+        UPDATE published_award_financial_assistance AS pafa
+        SET high_comp_officer1_full_na = duns.high_comp_officer1_full_na,
+            high_comp_officer1_amount = duns.high_comp_officer1_amount,
+            high_comp_officer2_full_na = duns.high_comp_officer2_full_na,
+            high_comp_officer2_amount = duns.high_comp_officer2_amount,
+            high_comp_officer3_full_na = duns.high_comp_officer3_full_na,
+            high_comp_officer3_amount = duns.high_comp_officer3_amount,
+            high_comp_officer4_full_na = duns.high_comp_officer4_full_na,
+            high_comp_officer4_amount = duns.high_comp_officer4_amount,
+            high_comp_officer5_full_na = duns.high_comp_officer5_full_na,
+            high_comp_officer5_amount = duns.high_comp_officer5_amount
+        FROM duns
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+            AND pafa.awardee_or_recipient_uniqu = duns.awardee_or_recipient_uniqu
+            AND duns.high_comp_officer1_full_na IS NOT NULL;
+    """
+    sess.execute(query.format(submission_id=submission_id))
+
+    log_derivation('Completed executive compensation derivation', submission_id, start_time)
 
 
-def derive_labels(obj):
+def derive_labels(sess, submission_id):
     """ Deriving labels for codes entered by the user
 
         Args:
-            obj: a dictionary containing the details we need to derive from and to
+            sess: the current DB session
+            submission_id: The ID of the submission derivations are being run for
     """
+    start_time = datetime.now()
+    log_derivation('Beginning label derivation', submission_id)
 
-    # Derive description for ActionType
-    if obj['action_type']:
-        obj['action_type_description'] = ACTION_TYPE_DICT.get(obj['action_type'].upper())
-    else:
-        obj['action_type_description'] = None
+    # Action type description derivation
+    action_type_values = '), ('.join('\'{}\', \'{}\''.format(name, desc) for name, desc in ACTION_TYPE_DICT.items())
+    query = """
+        WITH action_type_desc AS
+            (SELECT *
+            FROM (VALUES ({action_types})) as action_type_desc(letter, description))
+        UPDATE published_award_financial_assistance AS pafa
+        SET action_type_description = description
+        FROM action_type_desc AS atd
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+            AND atd.letter = UPPER(pafa.action_type);
+    """
+    sess.execute(query.format(submission_id=submission_id, action_types=action_type_values))
 
-    # Derive description for AssistanceType
-    if obj['assistance_type']:
-        obj['assistance_type_desc'] = ASSISTANCE_TYPE_DICT.get(obj['assistance_type'])
-    else:
-        obj['assistance_type_desc'] = None
+    # Assistance type description derivation
+    assistance_type_values = '), ('.join('\'{}\', \'{}\''.format(name, desc)
+                                         for name, desc in ASSISTANCE_TYPE_DICT.items())
+    query = """
+        WITH assistance_type_description AS
+            (SELECT *
+            FROM (VALUES ({assistance_types})) as assistance_type_description(letter, description))
+        UPDATE published_award_financial_assistance AS pafa
+        SET assistance_type_desc = description
+        FROM assistance_type_description AS atd
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+            AND atd.letter = UPPER(pafa.assistance_type);
+    """
+    sess.execute(query.format(submission_id=submission_id, assistance_types=assistance_type_values))
 
-    # Derive description for CorrectionDeleteIndicator
-    if obj['correction_delete_indicatr']:
-        obj['correction_delete_ind_desc'] = CORRECTION_DELETE_IND_DICT.get(obj['correction_delete_indicatr'].upper())
-    else:
-        obj['correction_delete_ind_desc'] = None
+    # CorrectionDeleteIndicator description derivation
+    cdi_values = '), ('.join('\'{}\', \'{}\''.format(name, desc) for name, desc in CORRECTION_DELETE_IND_DICT.items())
+    query = """
+        WITH cdi_desc AS
+            (SELECT *
+            FROM (VALUES ({cdi_types})) as cdi_desc(letter, description))
+        UPDATE published_award_financial_assistance AS pafa
+        SET correction_delete_ind_desc = description
+        FROM cdi_desc
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+            AND cdi_desc.letter = UPPER(pafa.correction_delete_indicatr);
+    """
+    sess.execute(query.format(submission_id=submission_id, cdi_types=cdi_values))
 
-    # Derive description for RecordType
-    if obj['record_type']:
-        obj['record_type_description'] = RECORD_TYPE_DICT.get(obj['record_type'])
-    else:
-        obj['record_type_description'] = None
+    # Record Type description derivation
+    record_type_values = '), ('.join('{}, \'{}\''.format(name, desc) for name, desc in RECORD_TYPE_DICT.items())
+    query = """
+        WITH record_type_desc AS
+            (SELECT *
+            FROM (VALUES ({record_types})) as record_type_desc(letter, description))
+        UPDATE published_award_financial_assistance AS pafa
+        SET record_type_description = description
+        FROM record_type_desc AS rtd
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+            AND rtd.letter = pafa.record_type;
+    """
+    sess.execute(query.format(submission_id=submission_id, record_types=record_type_values))
 
-    # Derive description for BusinessTypes
-    if obj['business_types']:
-        types_list = []
-        types_string = None
-        # loop through all the entries in business_types
-        for i in obj['business_types'].upper():
-            type_desc = BUSINESS_TYPE_DICT.get(i)
-            # If a valid business type was entered, append it to the list
-            if type_desc:
-                types_list.append(type_desc)
-        # If there was at least one valid business type, turn it into a string, separated by semicolons
-        if len(types_list) > 0:
-            types_string = ";".join(types_list)
-        obj['business_types_desc'] = types_string
-    else:
-        obj['business_types_desc'] = None
+    # Business Funds Indicator description derivation
+    business_funds_values = '), ('.join('\'{}\', \'{}\''.format(name, desc)
+                                        for name, desc in BUSINESS_FUNDS_IND_DICT.items())
+    query = """
+        WITH business_funds_ind_description AS
+            (SELECT *
+            FROM (VALUES ({business_funds_ind})) as business_funds_ind_description(letter, description))
+        UPDATE published_award_financial_assistance AS pafa
+        SET business_funds_ind_desc = description
+        FROM business_funds_ind_description AS bfid
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+            AND bfid.letter = UPPER(pafa.business_funds_indicator);
+    """
+    sess.execute(query.format(submission_id=submission_id, business_funds_ind=business_funds_values))
 
-    # Derive description for BusinessFundsIndicator
-    if obj['business_funds_indicator']:
-        obj['business_funds_ind_desc'] = BUSINESS_FUNDS_IND_DICT.get(obj['business_funds_indicator'].upper())
-    else:
-        obj['business_funds_ind_desc'] = None
+    # Business types description derivation
+    business_types_values = '), ('.join('\'{}\', \'{}\''.format(name, desc)
+                                        for name, desc in BUSINESS_TYPE_DICT.items())
+    query = """
+        WITH business_type_desc AS
+            (SELECT *
+            FROM(VALUES ({business_types})) as business_type_desc(letter, description)),
+        aggregated_business_types AS
+            (SELECT published_award_financial_assistance_id,
+                string_agg(btd.description, ';' order by ordinality) AS aggregated
+            FROM published_award_financial_assistance AS pafa,
+                unnest(string_to_array(pafa.business_types, NULL)) WITH ORDINALITY AS u(business_type_id, ordinality)
+            LEFT JOIN business_type_desc AS btd
+                ON btd.letter = UPPER(business_type_id)
+            WHERE submission_id = {submission_id}
+                AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D'
+            GROUP BY published_award_financial_assistance_id)
+        UPDATE published_award_financial_assistance AS pafa
+        SET business_types_desc = abt.aggregated
+        FROM aggregated_business_types AS abt
+        WHERE
+            abt.published_award_financial_assistance_id = pafa.published_award_financial_assistance_id;
+    """
+    sess.execute(query.format(submission_id=submission_id, business_types=business_types_values))
+
+    log_derivation('Completed label derivation', submission_id, start_time)
 
 
-def set_active(obj):
+def set_active(sess, submission_id):
     """ Setting active
 
         Args:
-            obj: a dictionary containing the details we need to derive from and to
+            sess: the current DB session
+            submission_id: The ID of the submission derivations are being run for
     """
-    if obj['correction_delete_indicatr'] and obj['correction_delete_indicatr'].upper() == 'D':
-        obj['is_active'] = False
-    else:
-        obj['is_active'] = True
+    start_time = datetime.now()
+    log_derivation('Beginning active derivation', submission_id)
+
+    query = """
+        UPDATE published_award_financial_assistance AS pafa
+        SET is_active = True
+        WHERE submission_id = {submission_id}
+            AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D';
+    """
+    sess.execute(query.format(submission_id=submission_id))
+
+    log_derivation('Completed active derivation', submission_id, start_time)
 
 
-def fabs_derivations(obj, sess, state_dict, country_dict, sub_tier_dict, cfda_dict, county_dict, office_dict,
-                     exec_comp_dict):
-    """ Performs derivations related to publishing a FABS record on a single row
+def set_modified_at(sess, submission_id):
+    """ Setting modified_at date
 
         Args:
-            obj: a dictionary containing the details we need to derive from and to
             sess: the current DB session
-            state_dict: a dictionary containing all the data for State objects keyed by state code
-            country_dict: a dictionary containing all the data for Country objects keyed by country code
-            sub_tier_dict: a dictionary containing all the data for SubTierAgency objects keyed by sub tier code
-            cfda_dict: a dictionary containing data for all CFDA objects keyed by cfda number
-            county_dict: a dictionary containing all the data for County objects keyed by state code + county number
-            office_dict: a dictionary containing all the data for Offices objects keyed by code
-            exec_comp_dict: a dictionary containing all the data for Executive Compensation data keyed by DUNS number
-
-        Returns:
-            The obj dictionary initially passed in but without the job_id or detached_award_financial_assistance_id
-            keys and with all the new keys created by derivations
+            submission_id: The ID of the submission derivations are being run for
     """
-    # copy log data and remove keys in the row left for logging
-    job_id = obj['job_id']
-    detached_award_financial_assistance_id = obj['detached_award_financial_assistance_id']
-    obj.pop('detached_award_financial_assistance_id', None)
-    obj.pop('job_id', None)
+    start_time = datetime.now()
+    log_derivation('Beginning modified_at derivation', submission_id)
 
-    # initializing a few of the derivations so the keys exist
-    obj['legal_entity_state_code'] = None
-    obj['legal_entity_state_name'] = None
-    obj['legal_entity_county_code'] = None
-    obj['legal_entity_county_name'] = None
-    obj['legal_entity_city_code'] = None
-    obj['legal_entity_city_name'] = None
-    obj['place_of_performance_zip5'] = None
-    obj['place_of_perform_zip_last4'] = None
-    obj['awarding_agency_code'] = None
-    obj['funding_agency_code'] = None
+    query = """
+        UPDATE published_award_financial_assistance AS pafa
+        SET modified_at = NOW()
+        WHERE submission_id = {submission_id};
+    """
+    sess.execute(query.format(submission_id=submission_id))
 
-    # Don't do any of these derivations on deletion records
-    if not obj['correction_delete_indicatr'] or obj['correction_delete_indicatr'].upper() != 'D':
-        # deriving total_funding_amount
-        federal_action_obligation = obj['federal_action_obligation'] or 0
-        non_federal_funding_amount = obj['non_federal_funding_amount'] or 0
-        obj['total_funding_amount'] = federal_action_obligation + non_federal_funding_amount
+    log_derivation('Completed modified_at derivation', submission_id, start_time)
 
-        derive_cfda(obj, cfda_dict, job_id, detached_award_financial_assistance_id)
 
-        derive_awarding_agency_data(obj, sub_tier_dict, office_dict)
+def fabs_derivations(sess, submission_id):
+    """ Performs derivations related to publishing a FABS submission
 
-        derive_funding_agency_data(obj, sub_tier_dict, office_dict)
+        Args:
+            sess: the current DB session
+            submission_id: The ID of the submission derivations are being run for
+    """
+    # TODO: Decide if we want to include the job in the logs
+    # TODO: Decide if we want to log each SQL query (including start/end or just the end) or just each function
+    derive_total_funding_amount(sess, submission_id)
 
-        ppop_code, ppop_state_code, ppop_state_name = derive_ppop_state(obj, state_dict)
+    derive_cfda(sess, submission_id)
 
-        derive_ppop_location_data(obj, sess, ppop_code, ppop_state_code, county_dict)
+    derive_awarding_agency_data(sess, submission_id)
 
-        derive_le_location_data(obj, sess, ppop_code, state_dict, ppop_state_code, ppop_state_name, county_dict)
+    derive_funding_agency_data(sess, submission_id)
 
-        derive_office_data(obj, office_dict, sess)
+    derive_ppop_state(sess, submission_id)
 
-        derive_le_city_code(obj, sess)
+    split_ppop_zip(sess, submission_id)
 
-        derive_ppop_country_name(obj, country_dict)
+    derive_ppop_location_data(sess, submission_id)
 
-        derive_le_country_name(obj, country_dict)
+    derive_le_location_data(sess, submission_id)
 
-        derive_ppop_code(obj)
+    derive_office_data(sess, submission_id)
 
-        derive_pii_redacted_ppop_data(obj)
+    derive_le_city_code(sess, submission_id)
 
-        split_ppop_zip(obj)
+    derive_ppop_country_name(sess, submission_id)
 
-        derive_parent_duns(obj, sess)
+    derive_le_country_name(sess, submission_id)
 
-        derive_executive_compensation(obj, exec_comp_dict)
+    derive_pii_redacted_ppop_data(sess, submission_id)
 
-        derive_labels(obj)
+    derive_parent_duns(sess, submission_id)
 
-        derive_ppop_scope(obj)
+    derive_executive_compensation(sess, submission_id)
 
-        # calculate business categories
-        obj['business_categories'] = get_business_categories(row=obj, data_type='fabs')
+    derive_labels(sess, submission_id)
 
-    set_active(obj)
+    derive_ppop_scope(sess, submission_id)
 
-    obj['modified_at'] = datetime.utcnow()
+    derive_fabs_business_categories(sess, submission_id)
 
-    return obj
+    set_active(sess, submission_id)
+
+    set_modified_at(sess, submission_id)
